requirements.txt
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
Django==5.2.6
h11==0.16.0
idna==3.10
numpy==2.3.3
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pycparser==2.23
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
selenium==4.35.0
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wheel==0.45.1
wsproto==1.2.0
